feat: add support for subscriptions
By now, the updated environment are sent through a `Channel` object,
of capacity 1 and dropping the contained value each time a new updated
environment is added. Also, as channels are implemented as consumables,
each time an item is sent and received, it is consumed and dropped when
received. This should be avoided in the context in which a subscription
is made by multiple clients, or the subscription is called multiple
times.